Add application embedding a ctkPythonConsole integrating PythonQt
The goal is to integrate IPython within the existing console and also
provide a way to interact using a notebook. Both the console and
the notebook should be available at the same time.

The console could either remain non IPython or also expose the IPython
interface.

This demo project has been successfully build on Ubuntu 13.10